==> R/1_Data collection.Rmd <==
---
title: "Data collection"
author: "Di Xiao"
date: "`r paste0('Initiated on 2021 April, compiled on ', format(Sys.time(), '%Y %b %d'))`"
output:
  html_document:
    code_folding: hide
    fig_height: 6
    fig_width: 6
    toc: yes
    number_sections: true
    toc_depth: 3
    toc_float: yes
    theme: paper
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# set up environment
```{r}
library(PhosR)
setwd("~/Dropbox (Sydney Uni)/Mining of publicly available phosphoproteomics data identify common signalome/repo/")
source("R/Functions.R")
```

# collect datasets from qPhos database
```{r}
datasets.qPhos <- read.delim("public resource/qPhos_all_data.txt", header = TRUE)

matDataset <- function(mat) {
    datasets <- list()
    Condition <- paste(sapply(mat$Sample, function(x)x), sapply(mat$Condition, function(x)x), sep = ";")
    mat$Condition <- Condition
    conditions <- unique(Condition)
    for (i in 1:length(conditions)) {    
        tmp <- mat[which(mat$Condition == conditions[i]),]
        Ids <- paste(paste(sapply(tmp[,2], function(x)paste(x)), 
                           paste(sapply(substr(tmp$Sequence.window, 8, 8), function(x)paste(x)), sapply(tmp$Position, function(x)paste(x)), sep = ""),
                           sep = ";"), ";", sep = "")
    Log2Ratio <- tmp[, "Log2Ratio"]
    datasets[[i]] <- as.matrix(sapply(split(Log2Ratio, Ids), mean))
    names(datasets)[i] <- paste(tmp[1,"Condition"], sep = ";")
    }
    return(datasets)
}

datasets <- matDataset(datasets.qPhos)

# adding more dataset from the updated qPhos basebase
path = "datasets/qPhos"
files <- list.files(path, pattern = "PMID", full.names = TRUE)

combine.datasets <- datasets
matAdd <- list()
for (i in 1:length(files)) {
  dataset.add <- read.delim(files[i], header = TRUE)
  matAdd[[i]] <- matDataset(dataset.add)
  combine.datasets <- c(combine.datasets, matAdd[[i]])
}

# save(combine.datasets, file = "datasets/datasets.qPhos.Rda")
```

# select dataset with more then 5000 phosphosite and reform the format to PhosphoExperimen
```{r}
HeLa_S3.pe <- mat(c(34:39))

HEK293.pe <- mat(c(43:46))

HUES9.pe <- mat(c(55:62))

Hela.MLN8054.pe <- mat(c(80:82))

HeLa.BI2536.pe <- mat(c(83:84))

GTL16.pe <- mat(c(88:91))

KG1.pe <- mat(c(94:96))

hTERT.RPE1.pe <- mat(c(97:98))

breast.cancer.pe <- mat(c(121:122))

SCC9.pe <- mat(c(146:149))

BOEC.Thrombin.pe <- mat(c(206:209))

Jurkat.PGE2.pe <- mat(c(222:226))

TIG3.pe <- mat(c(231:236))

A375.pe <- mat(c(242:244))

HaCaT.TGFB.pe <- mat(c(249:252))

MCF7.pe <- mat(c(281:291))

WM239A.pe <- mat(c(292:294))

A431.pe <- mat(c(417:418))

K562.pe <- mat(c(18:20))

Hela.EGF.pe <- mat(c(21:25))

A498.pe <- mat(c(40:42))

HEK293.Nocodazole.pe <- mat(c(100:101))

HEK293.PMA.pe <- mat(c(245:247))

MCF10A.pe <- mat(c(262:265))

NCI.pe <- mat(c(274:277))

PaSC.pe <- mat(c(311:312))

NCI_H929.pe <- mat(c(318:319))

Jurkat_E6.1.pe <- mat(c(336:339))

J.gamma1.pe <- mat(c(340:343))

U2OS.pe <- mat(c(344:346))

MDA.MB.231.pe <- mat(c(351:352))

HNSCC.pe <- mat(c(358:363))

HeLa.Flavopiridol.pe <- mat(c(367:368))

Daudi.pe <- mat(c(369:371))

DG.75.pe <- mat(c(376:377))

HEK293.Insulin.pe <- mat(c(437:439))

A431.K8.pe <- mat(c(440:443))

HFF.pe <- mat(c(444:448))

COLO205.pe <- mat(c(449:461))

HCT116.pe <- mat(c(462:474))

BT474.pe <- mat(c(475:476))

NCI.H1299.pe <- mat(c(495:496))

Jurkat.HIV.pe <- mat(c(497:501))

HeLa.Pervanadate.pe <- mat(c(510:511))

A204.pe <- mat(c(516:517))

MDA.MB.231.pe <- mat(c(548:549))

U2OS.etoposide.pe <- mat(c(555:556))

SH.SY5Y.pe <- mat(c(561:563))

A549.TNF.pe <- mat(c(566:568))

NB1.pe <- mat(c(592:594))

HCT116.pe <- mat(c(600:602))

BT474.pe <- mat(c(614:616))

HeLa.H2O2.pe <- mat(c(620:622)) 

phosData <- list(HeLa_S3.pe, HEK293.pe, HUES9.pe, Hela.MLN8054.pe, HeLa.BI2536.pe, GTL16.pe, KG1.pe, hTERT.RPE1.pe,
                 breast.cancer.pe, SCC9.pe, BOEC.Thrombin.pe, Jurkat.PGE2.pe, TIG3.pe, A375.pe, HaCaT.TGFB.pe, 
                 MCF7.pe, WM239A.pe, A431.pe, K562.pe, Hela.EGF.pe, A498.pe, HEK293.Nocodazole.pe, HEK293.PMA.pe, 
                 MCF10A.pe, NCI.pe, PaSC.pe, NCI_H929.pe, Jurkat_E6.1.pe, J.gamma1.pe, U2OS.pe, MDA.MB.231.pe,
                 HNSCC.pe, HeLa.Flavopiridol.pe, Daudi.pe, DG.75.pe, HEK293.Insulin.pe, A431.K8.pe, HFF.pe, COLO205.pe, 
                 HCT116.pe, BT474.pe, NCI_H929.pe, Jurkat.HIV.pe, HeLa.Pervanadate.pe, A204.pe, MDA.MB.231.pe,
                 U2OS.etoposide.pe, SH.SY5Y.pe, A549.TNF.pe, NB1.pe, HCT116.pe, BT474.pe, HeLa.H2O2.pe)


# save(phosData, file = "datasets/phosData_53_human.RData")

```
