==> R/5_Characterise SPSs_functions.Rmd <==
---
title: "Evaluating the stability of SPSs"
author: "Di Xiao"
date: "`r paste0('Initiated on 2021 April, compiled on ', format(Sys.time(), '%Y %b %d'))`"
output:
  html_document:
    code_folding: hide
    fig_height: 6
    fig_width: 6
    toc: yes
    number_sections: true
    toc_depth: 3
    toc_float: yes
    theme: paper
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# set environment
```{r}
library(PhosR)
library(ggplot2)
library(qusage)
library(org.Hs.eg.db)
library(clusterProfiler)
library(enrichplot)
library(ReactomePA)

load("../datasets/mid_bot_ran.RData")
load("../datasets/human_SPS_v2.RData")
hSPSs <- names(fisher.adj.p[which(fisher.adj.p < 0.01)])
SPSs.gene <- sapply(strsplit(hSPSs, ";"), function(x)x[1])
load("../datasets/feature_matrix.RData")
```

# kinase enrichment
```{r}
e1 <- pathwayOverrepresent(hSPSs, universe = phosphosites.human$Var1, alter = "greater", annotation = PhosphoSite.human)
e2 <- pathwayOverrepresent(ranSites[[6]], universe = phosphosites.human$Var1, alter = "greater", annotation = PhosphoSite.human)

## top 5
df1 <- data.frame(value = -log10(as.numeric(e1[1:5, 1])), names = rownames(e1[1:5,]), group <- rep("SPS", 5))
colnames(df1)[3] <- "group"
df2 <- data.frame(value = -log10(as.numeric(e2[which(rownames(e2) %in% rownames(e1)[1:5]), 1])), names = rownames(e2[which(rownames(e2) %in% rownames(e1)[1:5]),]), group <- rep("Random", 5))
colnames(df2)[3] <- "group"

df <- rbind(df1, df2)
df$names <- factor(df$names, levels = rownames(e1)[1:5])
df$group <- factor(df$group, levels = c("SPS", "Random"))

ggplot(df, aes(fill=group, y=value, x=names)) + 
    geom_bar(position="dodge", stat="identity") +
    scale_x_discrete(guide = guide_axis(angle = 45)) +
    theme_classic()

## all
df1 <- data.frame(value = -log10(as.numeric(e1[, 1])), names = rownames(e1[,]), group <- rep("SPS", nrow(e1)))
df1 <- df1[which(df1$value > 0), ]
colnames(df1)[3] <- "group"
df2 <- data.frame(value = -log10(as.numeric(e2[which(rownames(e2) %in% rownames(e1)), 1])), names = rownames(e2[which(rownames(e2) %in% rownames(e1)),]), group <- rep("Random", nrow(e2)))
df2 <- df2[which(df2$value > 0), ]
colnames(df2)[3] <- "group"

o <- union(df1$names, df2$names)
df1 <- rbind(df1, data.frame(value = 0, names = setdiff(o, df1$names), group = "SPS"))
df2 <- rbind(df2, data.frame(value = 0, names = setdiff(o, df2$names), group = "Random"))

df <- rbind(df1, df2)
df$names <- factor(df$names, levels = rownames(e1))
df$group <- factor(df$group, levels = c("SPS", "Random"))

ggplot(df, aes(fill=group, y=value, x=names)) + 
    geom_bar(position="dodge", stat="identity") +
    scale_x_discrete(guide = guide_axis(angle = 90)) +
    scale_fill_manual(values=c("#E41A1C", "#BC9DCC")) +
    theme_classic()
```

## phosphatase enrichment
```{r}
phosphatase <- read.delim("../public resource/Kinase_Phosphatase_substrates.txt")
phosphatase <- phosphatase[which(phosphatase$Enzyme.type == "Phosphatase"), ]
idx <- grep("\\?", phosphatase$Substrate.residue..amino.acid.type...position.)
phosphatase.sites <- phosphatase[-idx, ]
phosphatase.sites$Substrate.residue..amino.acid.type...position. <- gsub("Ser", "S", phosphatase.sites$Substrate.residue..amino.acid.type...position.)
phosphatase.sites$Substrate.residue..amino.acid.type...position. <- gsub("Thr", "T", phosphatase.sites$Substrate.residue..amino.acid.type...position.)
phosphatase.sites$Substrate.residue..amino.acid.type...position. <- gsub("Tyr", "Y", phosphatase.sites$Substrate.residue..amino.acid.type...position.)

phosphatase.annotation <- list()
for (phosphatase in unique(phosphatase.sites$Enzyme.Gene.Symbol)) {
    df <- phosphatase.sites[which(phosphatase.sites$Enzyme.Gene.Symbol == phosphatase), ]
    substrates <- paste(df$Substrate.Gene.Symbol, ";", sapply(strsplit(df$Substrate.residue..amino.acid.type...position.,"-"), function(x) paste(x[1], x[2], sep = "")), ";", sep = "")
    phosphatase.annotation[[phosphatase]] <- substrates
}

phosphatase.genes <- list()
for (phosphatase in unique(phosphatase.sites$Enzyme.Gene.Symbol)) {
    df <- phosphatase.sites[which(phosphatase.sites$Enzyme.Gene.Symbol == phosphatase), ]
    substrates <- df$Substrate.Gene.Symbol
    phosphatase.genes[[phosphatase]] <- substrates
}

## enrichment (nothing enriched for hSPSs on sites level)
e1 <- pathwayOverrepresent(hSPSs, universe = phosphosites.human$Var1, alter = "greater", annotation = phosphatase.sites)
e2 <- pathwayOverrepresent(ranSites[[6]], universe = phosphosites.human$Var1, alter = "greater", annotation = phosphatase.sites)

## phosphates enrichment on gene level
e1 <- pathwayOverrepresent(SPSs.gene, universe = phosphosites.human$Var1, alter = "greater", annotation = phosphatase.genes)
e2 <- pathwayOverrepresent(sapply(strsplit(as.character(ranSites[[6]]), ";"), function(x)x[1]), 
                           universe = sapply(strsplit(as.character(phosphosites.human$Var1), ";"), function(x)x[1]),
                           alter = "greater", annotation = phosphatase.genes)

df1 <- data.frame(value = -log10(as.numeric(e1[1:5, 1])), names = rownames(e1[1:5,]), group <- rep("SPS", 5))
colnames(df1)[3] <- "group"
df2 <- data.frame(value = -log10(as.numeric(e2[which(rownames(e2) %in% rownames(e1)[1:5]), 1])), names = rownames(e2[which(rownames(e2) %in% rownames(e1)[1:5]),]), group <- rep("Random", 5))
colnames(df2)[3] <- "group"

df <- rbind(df1, df2)
df$names <- factor(df$names, levels = rownames(e1)[1:5])
df$group <- factor(df$group, levels = c("SPS", "Random"))

ggplot(df, aes(fill=group, y=value, x=names)) + 
    geom_bar(position="dodge", stat="identity") +
    scale_x_discrete(guide = guide_axis(angle = 45)) +
    theme_classic()
```


## GO enrichment
```{r}
## BP enrithment
SPSs.entrze <- unname(mapIds(org.Hs.eg.db, SPSs.gene, 'ENTREZID', 'SYMBOL'))

go_enrichBP_SPS <- enrichGO(gene = unique(SPSs.gene),
                      OrgDb = org.Hs.eg.db, 
                      keyType = 'SYMBOL',
                      ont = "BP",
                      pvalueCutoff = 0.05, 
                      qvalueCutoff = 0.10)
SPS.BP <- data.frame(term = go_enrichBP_SPS@result$Description[1:5], pvalue = -log10(go_enrichBP_SPS@result$p.adjust[1:5]),
                     group = "SPS")

Random.gene <- sapply(strsplit(as.character(ranSites[[6]]), ";"), function(x)x[1])
Random.entrze <- unname(mapIds(org.Hs.eg.db, Random.gene, 'ENTREZID', 'SYMBOL'))

go_enrichBP_Random <- enrichGO(gene = unique(Random.gene),
                      OrgDb = org.Hs.eg.db, 
                      keyType = 'SYMBOL',
                      ont = "BP",
                      pvalueCutoff = 0.05, 
                      qvalueCutoff = 0.10)
Random.BP <- data.frame(term = go_enrichBP_Random@result$Description[which(go_enrichBP_Random@result$Description %in% go_enrichBP_SPS@result$Description[1:5])], pvalue = -log10(go_enrichBP_Random@result$p.adjust[which(go_enrichBP_Random@result$Description %in% go_enrichBP_SPS@result$Description[1:5])]), group = "Random")

df <- rbind(SPS.BP, Random.BP)
df$term <- factor(df$term, levels = go_enrichBP_SPS@result$Description[1:5])
df$group <- factor(df$group, levels = c("SPS", "Random"))

ggplot(df, aes(fill=group, y=pvalue, x=term)) + 
    geom_bar(position="dodge", stat="identity") +
    scale_x_discrete(guide = guide_axis(angle = 45)) +
    theme_classic()

## CC enrichment
go_enrichCC_SPS <- enrichGO(gene = unique(SPSs.gene),
                      OrgDb = org.Hs.eg.db, 
                      keyType = 'SYMBOL',
                      ont = "CC",
                      pvalueCutoff = 0.05, 
                      qvalueCutoff = 0.10)

SPS.CC <- data.frame(term = go_enrichCC_SPS@result$Description[1:5], pvalue = -log10(go_enrichCC_SPS@result$p.adjust[1:5]),
                     group = "SPS")

go_enrichCC_Random <- enrichGO(gene = unique(Random.gene),
                      OrgDb = org.Hs.eg.db, 
                      keyType = 'SYMBOL',
                      ont = "CC",
                      pvalueCutoff = 0.05, 
                      qvalueCutoff = 0.10)
Random.CC <- data.frame(term = go_enrichCC_Random@result$Description[which(go_enrichCC_Random@result$Description %in% go_enrichCC_SPS@result$Description[1:5])], pvalue = -log10(go_enrichCC_Random@result$p.adjust[which(go_enrichCC_Random@result$Description %in% go_enrichCC_SPS@result$Description[1:5])]), group = "Random")

df <- rbind(Random.CC, SPS.CC)
df$term <- factor(df$term, levels = go_enrichCC_SPS@result$Description[1:5])
df$group <- factor(df$group, levels = c("SPS", "Random"))

ggplot(df, aes(fill=group, y=pvalue, x=term)) + 
    geom_bar(position="dodge", stat="identity") +
    scale_x_discrete(guide = guide_axis(angle = 45)) +
    theme_classic()

## MF enrichment
go_enrichMF_SPS <- enrichGO(gene = unique(SPSs.gene),
                      OrgDb = org.Hs.eg.db, 
                      keyType = 'SYMBOL',
                      ont = "MF",
                      pvalueCutoff = 0.05, 
                      qvalueCutoff = 0.10)

SPS.MF <- data.frame(term = go_enrichMF_SPS@result$Description[1:5], pvalue = -log10(go_enrichMF_SPS@result$p.adjust[1:5]),
                     group = rep("SPS", 5))

go_enrichMF_Random <- enrichGO(gene = unique(Random.gene),
                      OrgDb = org.Hs.eg.db, 
                      keyType = 'SYMBOL',
                      ont = "MF",
                      pvalueCutoff = 0.05, 
                      qvalueCutoff = 0.10)
Random.MF <- data.frame(term = go_enrichMF_Random@result$Description[which(go_enrichMF_Random@result$Description %in% go_enrichMF_SPS@result$Description[1:5])], pvalue = -log10(go_enrichMF_Random@result$p.adjust[which(go_enrichMF_Random@result$Description %in% go_enrichMF_SPS@result$Description[1:5])]), group = "Random")

df <- rbind(Random.MF, SPS.MF)
df$term <- factor(df$term, levels = go_enrichMF_SPS@result$Description[1:5])
df$group <- factor(df$group, levels = c("SPS", "Random"))

ggplot(df, aes(fill=group, y=pvalue, x=term)) + 
    geom_bar(position="dodge", stat="identity") +
    scale_x_discrete(guide = guide_axis(angle = 45)) +
    theme_classic()

```

## reactome pathway enrichment
```{r}
load("../datasets/Pathways.rda")
e1 <- pathwayOverrepresent(SPSs.gene, annotation= Pathways.reactome, universe = unique(phosphosites.human$geneSymbol), alter = "greater")
e2 <- pathwayOverrepresent(Random.gene, annotation= Pathways.reactome, universe = unique(phosphosites.human$geneSymbol), alter = "greater")

df1 <- data.frame(value = -log10(as.numeric(e1[1:5, 1])), names = rownames(e1[1:5,]), group <- rep("SPS", 5))
colnames(df1)[3] <- "group"
df2 <- data.frame(value = -log10(as.numeric(e2[which(rownames(e2) %in% rownames(e1)[1:5]), 1])), names = rownames(e2[which(rownames(e2) %in% rownames(e1)[1:5]),]), group <- rep("Random", 5))
colnames(df2)[3] <- "group"

df <- rbind(df1, df2)
df$names <- factor(df$names, levels = rownames(e1)[1:5])
df$group <- factor(df$group, levels = c("SPS", "Random"))

ggplot(df, aes(fill=group, y=value, x=names)) + 
    geom_bar(position="dodge", stat="identity") +
    scale_x_discrete(guide = guide_axis(angle = 45)) +
    theme_classic()
```

## Biocarta enrichment
```{r}
Biocarta.annotation <- read.gmt("../public resource/c2.cp.biocarta.v7.4.symbols.gmt")

e1 <- pathwayOverrepresent(SPSs.gene, 
                           universe = unique(phosphosites.human$geneSymbol), 
                           alter = "greater", annotation = Biocarta.annotation)
e2 <- pathwayOverrepresent(Random.gene, 
                           universe = unique(phosphosites.human$geneSymbol),
                           alter = "greater", annotation = Biocarta.annotation)

df1 <- data.frame(value = -log10(as.numeric(e1[1:5, 1])), names = rownames(e1[1:5,]), group <- rep("SPS", 5))
colnames(df1)[3] <- "group"
df2 <- data.frame(value = -log10(as.numeric(e2[which(rownames(e2) %in% rownames(e1)[1:5]), 1])), names = rownames(e2[which(rownames(e2) %in% rownames(e1)[1:5]),]), group <- rep("Random", 5))
colnames(df2)[3] <- "group"

df <- rbind(df1, df2)
df$names <- factor(df$names, levels = rownames(e1)[1:5])
df$group <- factor(df$group, levels = c("SPS", "Random"))

ggplot(df, aes(fill=group, y=value, x=names)) + 
    geom_bar(position="dodge", stat="identity") +
    scale_x_discrete(guide = guide_axis(angle = 45)) +
    theme_classic()

```



## KEGG enrichment
```{r}
KEGG.annotation <- read.gmt("../public resource/c2.cp.kegg.v7.4.symbols.gmt")

e1 <- pathwayOverrepresent(SPSs.gene, 
                           universe = phosphosites.human$geneSymbol,
                           alter = "greater", annotation = KEGG.annotation)
e2 <- pathwayOverrepresent(Random.gene, 
                           universe = phosphosites.human$geneSymbol,
                           alter = "greater", annotation = KEGG.annotation)

df1 <- data.frame(value = -log10(as.numeric(e1[1:5, 1])), names = rownames(e1[1:5,]), group <- rep("SPS", 5))
colnames(df1)[3] <- "group"
df2 <- data.frame(value = -log10(as.numeric(e2[which(rownames(e2) %in% rownames(e1)[1:5]), 1])), names = rownames(e2[which(rownames(e2) %in% rownames(e1)[1:5]),]), group <- rep("Random", 5))
colnames(df2)[3] <- "group"

df <- rbind(df1, df2)
df$names <- factor(df$names, levels = rownames(e1)[1:5])
df$group <- factor(df$group, levels = c("SPS", "Random"))

ggplot(df, aes(fill=group, y=value, x=names)) + 
    geom_bar(position="dodge", stat="identity") +
    scale_x_discrete(guide = guide_axis(angle = 45)) +
    theme_classic()
```
