==> R/7_Reproducibility.Rmd <==
---
title: "7_Reproducibility"
author: "Di Xiao"
date: "`r paste0('Initiated on 2021 Nov, compiled on ', format(Sys.time(), '%Y %b %d'))`"
output:
  html_document:
    code_folding: hide
    fig_height: 6
    fig_width: 6
    toc: yes
    number_sections: true
    toc_depth: 3
    toc_float: yes
    theme: paper
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(PhosR)
source("R/Functions.R")
```

## datasets
```{r}
## PMID: 31332097
hESC <- read.delim2("datasets/PMID31332097.txt")
sites <- paste(toupper(hESC$Gene.names), ";",
               paste(substr(hESC$Sequence.window, 16, 16), sapply(strsplit(hESC$Positions.within.proteins, ";"), function(x)x[1]), sep = ""),
               ";", sep = "")
hESC <- hESC[, 8:ncol(hESC)]
hESC <- hESC[-4932, ]
hESC <- apply(hESC, 2, as.numeric)
hESC.imputed <- PhosR::tImpute(hESC)
rownames(hESC.imputed) <- sites[-4932]
grps <- gsub("_R\\d$", "", colnames(hESC.imputed))
hESC.mean <- meanAbundance(hESC.imputed, grps = grps)
hESC.mean <- hESC.mean[,2:ncol(hESC.mean)] - hESC.mean[,1]

## PMID: 33947818
HCT116 <- readxl::read_xlsx("datasets/PMID33947818.xlsx", sheet = 1)
sites <- paste(sapply(strsplit(HCT116$Gene.names, ";"), function(x)x[1]), 
            ";", HCT116$Positions.within.proteins, ";", sep = "")
idx <- which(duplicated(sites))
HCT116 <- HCT116[-idx, c(3,5)]
rownames(HCT116) <- sites[-idx]

load("datasets/feature_matrix.RData")
matched <- match(rownames(HCT116), phosphosites.human$siteNoResidue)
sites.matched <- phosphosites.human$Var1[matched]

idx <- which(is.na(sites.matched))
HCT116 <- HCT116[-idx, ]
HCT116 <- apply(HCT116, 2, as.numeric)
rownames(HCT116) <- sites.matched[-idx]
HCT116.mean <- as.matrix(HCT116)

## PMID33357429
Tcells <- readxl::read_xlsx("datasets/PMID33357429.xlsx", sheet = 2)
sites <- paste(sapply(strsplit(Tcells$`Gene names`, ";"), function(x)x[1]), 
                ";",
                paste(Tcells$`Amino acid`, sapply(strsplit(Tcells$`Positions within proteins`, ";"), function(x)x[1]), sep = ""),
               ";", sep = "")

idx <- which(duplicated(sites))
Tcells <- Tcells[-idx, grep("^log2 average", colnames(Tcells))]
Tcells <- apply(Tcells, 2, as.numeric)
rownames(Tcells) <- sites[-idx]
Tcells.mean <- as.matrix(Tcells)

## PMID32154729  
AGS <- readxl::read_xlsx("datasets/PMID32154729 .xlsx", sheet = 2, skip = 1)
sites <- paste(AGS$`Gene name`, ";",
               paste(AGS$`Phosphorylated residue`, AGS$`Phosphosite position in proteins`, sep = ""),
               ";", sep = "")
AGS <- AGS[, grep("^Ratio H/L normalized", colnames(AGS))]
rownames(AGS) <- sites
AGS <- apply(AGS, 2, as.numeric)
AGS.imputed <- tImpute(AGS)
rownames(AGS.imputed) <- sites
AGS.mean <- log2(AGS.imputed)


load("datasets/datasets.qPhos.Rda")
HeLa.pe <- mat(c(2:5))
HeLa <- HeLa.pe@assays@data$normalised

HeLa.2.pe <- mat(c(22:23))
HeLa.2 <- HeLa.2.pe@assays@data$normalised

AML.pe <- mat(c(155:156))
AML <- AML.pe@assays@data$normalised

Kasumi_1.pe <- mat(c(158:162))
Kasumi_1 <- Kasumi_1.pe@assays@data$normalised

p31.pe <- mat(c(163:168))
p31 <- p31.pe@assays@data$normalised

MCF.pe <- mat(c(193:196))
MCF <- MCF.pe@assays@data$normalised

DLD.pe <- mat(c(295:296))
DLD <- DLD.pe@assays@data$normalised

DFC.pe <- mat(c(215:216))
DFC <- DFC.pe@assays@data$normalised

HeLa.pe <- mat(c(298:299))
HeLa.Kyoto <- HeLa.pe@assays@data$normalised

Hep_G2 <- mat(c(306:307))
Hep_G2 <- Hep_G2@assays@data$normalised

THP_1 <- mat(c(308:310))
THP_1 <- THP_1@assays@data$normalised

liver <- mat(c(353, 354))
liver <- liver@assays@data$normalised

Melanocytes <- mat(c(380, 381))
Melanocytes <- Melanocytes@assays@data$normalised

SW48 <- mat(c(390:395))
SW48 <- SW48@assays@data$normalised

blood <- mat(c(406:409))
blood <- blood@assays@data$normalised

NHA <- mat(c(419:420))
NHA <- NHA@assays@data$normalised

blood.ADP <- mat(c(426:430))
blood.ADP <- blood.ADP@assays@data$normalised

HT22 <- mat(c(491:493))
HT22 <- HT22@assays@data$normalised

ARPE_19 <- mat(c(520:524))
ARPE_19 <- ARPE_19@assays@data$normalised

U2OS <- mat(c(525:527))
U2OS <- U2OS@assays@data$normalised

NB4 <- mat(c(528:530))
NB4 <- NB4@assays@data$normalised

CD4T <- mat(c(578:583))
CD4T <- CD4T@assays@data$normalised
```

## Run framework
```{r}
datasets <- list(HCT116.mean, Tcells.mean, AGS.mean, hESC.mean, 
                 HeLa, HeLa.2, AML, Kasumi_1, hESC.mean, p31,
                 MCF, DLD, DFC, HeLa.Kyoto, Hep_G2,
                 THP_1, liver, Melanocytes, SW48, blood,
                 NHA, blood.ADP, HT22, ARPE_19, U2OS,
                 NB4, CD4T)
sites <- sapply(datasets, rownames)
sites.max <- list()
for (i in 1:length(datasets)) {
    dataset <- datasets[[i]]
    sites.max[[i]] <- apply(dataset, 1, function(x)(max(abs(x))))
}
mat.max <- sapply(sites.max, function(x) sort(abs(x), decreasing = TRUE))

o <- as.data.frame(table(unlist(sites)))
o.order <- o[order(o$Freq, decreasing = TRUE), ]

Ts <- data.frame(mat.max[[1]][as.character(o$Var1)])
for (i in 2:length(datasets)) {
    Ts <- cbind(Ts, mat.max[[i]][as.character(o$Var1)])
}
rownames(Ts) <- as.character(o$Var1)
Ts <- Ts[-(1:40),]

maxFC.mat <- Ts
recurrence <- o.order
result <- list(recurrence, maxFC.mat)

SPSs.r <- identifySPSs(result[[1]], result[[2]],  n=n)

load("datasets/human_SPS_v2.Rdata")
hSPSs <- names(fisher.adj.p[which(fisher.adj.p < 0.05)])

o <- intersect(names(SPSs.r[[3]]), names(fisher.adj.p))
mat.r <- data.frame(SPSs.r[[3]][o], fisher.adj.p[o])
mat.quan <- normalize.quantiles(as.matrix(mat.r))

d <- densCols(mat.quan[,1], mat.quan[,2], colramp=colorRampPalette(c("black", "white")))
cols <-  colorRampPalette(c("gray", "darkblue", "deepskyblue2", "deepskyblue"))(256)
colss <- cols[col2rgb(d)[1,] + 1L]
plot(-log10(mat.quan[,1]), -log10(mat.quan[,2]), col=colss)
cor.test(-log10(mat.quan[,1]), -log10(mat.quan[,2]))
```
